third_party/WebKit/LayoutTests/http/tests/devtools/sources/autocomplete-hide-on-smart-brace.html

<html>
<head>
<script src="../sources/debugger/resources/edit-me.js"></script>
<script src="../resources/editor-test.js"></script>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/debugger-test.js"></script>
<script>

function test() {
  SourcesTestRunner.showScriptSource('edit-me.js', onSourceFrame);

  var textEditor;
  function onSourceFrame(sourceFrame) {
    textEditor = sourceFrame.textEditor;
    textEditor.element.focus();
    TestRunner.runTestSuite(testSuite);
  }

  var testSuite = [
    function testSummonSuggestBox(next) {
      TestRunner.addSniffer(
          TextEditor.TextEditorAutocompleteController.prototype, '_onSuggestionsShownForTest', onSuggestionsShown);

      textEditor.setText('one\n()');
      textEditor.setSelection(TextUtils.TextRange.createFromLocation(1, 1));
      SourcesTestRunner.typeIn(textEditor, 'o', function() {});

      function onSuggestionsShown() {
        TestRunner.addResult('Suggestions shown.');
        next();
      }
    },

    function testTypeSmartBrace(next) {
      TestRunner.addSniffer(
          TextEditor.TextEditorAutocompleteController.prototype, '_onSuggestionsHiddenForTest', onSuggestionsHidden);
      SourcesTestRunner.typeIn(textEditor, ')', function() {});

      function onSuggestionsHidden() {
        TestRunner.addResult('Suggestions hidden.');
        next();
      }
    },
  ];
}

</script>
</head>

<body onload="runTest();">
<p>
Verify that suggest box gets hidden whenever a cursor jumps over smart brace.
</p>
</body>
</html>
