third_party/WebKit/LayoutTests/http/tests/devtools/sources/autocomplete-hide-on-smart-brace-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verify that suggest box gets hidden whenever a cursor jumps over smart brace.


Running: testSummonSuggestBox
Suggestions shown.

Running: testTypeSmartBrace
Suggestions hidden.
